requirements.txt
SQLAlchemy~=1.4.45
python-dotenv~=0.21.0
python-telegram-bot~=13.15
